% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertItem.R
\name{invertItem}
\alias{invertItem}
\title{Invert item scoring}
\usage{
invertItem(x, min, max)
}
\arguments{
\item{x}{A vector to be inverted}

\item{min}{The minimum possible score for this vector}

\item{max}{The maximum possible score for this vector}
}
\value{
A vector of numbers
}
\description{
A function to invert the scoring of a variable (e.g. a likert item)
}
\examples{
x <- 1:7
y <- -4:4
z <- 3:21
invertItem(x, 1, 7)
invertItem(y, -4, 4)
invertItem(z, 3, 21)
}
